lesson_9/requirements.txt
Jinja2==3.1.2
MarkupSafe==2.1.2
parse==1.19.0
waitress==2.1.2
WebOb==1.8.7

requests==2.28.2
pytest==7.2.1attrs==22.2.0
certifi==2022.12.7
charset-normalizer==3.0.1
colorama==0.4.6
exceptiongroup==1.1.0
idna==3.4
iniconfig==2.0.0
Jinja2==3.1.2
MarkupSafe==2.1.2
packaging==23.0
parse==1.19.0
pluggy==1.0.0
pytest==7.2.1
requests==2.28.2
requests-wsgi-adapter==0.4.1
tomli==2.0.1
urllib3==1.26.14
waitress==2.1.2
WebOb==1.8.7
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==3.0.1
colorama==0.4.6
exceptiongroup==1.1.0
idna==3.4
iniconfig==2.0.0
Jinja2==3.1.2
MarkupSafe==2.1.2
packaging==23.0
parse==1.19.0
pluggy==1.0.0
pytest==7.2.1
requests==2.28.2
requests-wsgi-adapter==0.4.1
tomli==2.0.1
urllib3==1.26.14
waitress==2.1.2
WebOb==1.8.7
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==3.0.1
colorama==0.4.6
exceptiongroup==1.1.0
idna==3.4
iniconfig==2.0.0
Jinja2==3.1.2
MarkupSafe==2.1.2
packaging==23.0
parse==1.19.0
pluggy==1.0.0
pytest==7.2.1
requests==2.28.2
requests-wsgi-adapter==0.4.1
tomli==2.0.1
urllib3==1.26.14
waitress==2.1.2
WebOb==1.8.7

jsonpickle==3.0.1